tests/SConscript added a library that wasn't used, and errored out when declared so I removed it.
src/events/Events.cc had a variable that was uninitialized;  under gcc 3.x,
this was initialized to 0 by default (so everything worked);  under gcc 4.x,
this was left uninitialized so things were thrown off.   The fix was to 
explicitedly declare _turnEventsOff to false, which it should have been in
the first place.