added a comment to notify testers that if this test fails to try and remove the /tmp/test_events_3 file if it exists.
The problem is that when a local UNIX domain socket is created and  
it's owned by someone else, a new person can't create a socket with  
the same name.  This is a well known problem with UNIX domain  
sockets.   The socket itself looks like a file in the filesystem.   
Removing that file fixes the problem.

This has a trac ticket assigned.