Adds functionality for ticket #841 to allow the broker to filter messages based on certain conditions.
JMS allows messages to be filtered according
to the contents of the message properties.
Basic fields are int/float/boolean/string.

LSST Events payload (the PropertySet) is currently
serialized and sent as a text string.   We do
this to retain the data integrity of the PropertySet
which would be lost for some types (high precision
floats, "long long", for example).  If we put these
in the JMS properties, the data would be corrupted.

Therefore, an additional set of properties, which only 
needs to be specified at the time of EventTransmitter 
construction, has been added.  This permits the filtering
info to only be necessary to be set only once, instead
of grabbing this information from every single PropertySet
that is passed.   This also permits the later "users"
of the Transmitter (which get it through
a call to EventSystem) to not have to be aware which
fields are being added (runid, for example).

On the Receiver side, messages are filtered through
the selector syntax of JMS.  This lets us specify
SQL92-style expressions to do filtering.  A
receiver can specify:  "runid = 'srp2009'" to
only receive messages whose properties match that. 
Much more complex expressions are possible.

I've long had on my "to do" list to serialize only
the data incompatible parts of a DataProperty (now
PropertySet), and add the rest as JMS properties so
they can be used in these SQL92 expressions.  This
is a good start.


EventTransmitter.cc/.h:

Adds a new constructor method which allows the
user to pass a PropertySet which will be added
as properties to the Event.

EventSelector.cc/.h:

Added a new constructor method which allows the
user to pass a JMS message selector, which is
sent to the broker to have it filter messages
it sends to this receiver.

tests/Events_8.py

added a unit test that demonstrates this capability.