Added ability to specify filterable keywords.
Event (and their subclasses) carry certain standard keywords in message headers that are filterable server-side, when used in conjunction with selectors.   The rest of the property set information is kept in the payload of the message, and is not filterable with selectors.

This change allows the user to add keywords and values that will be added into the message header, so they can be filtered server-side as well.