Removed creation and deconstruction of OriginatorIDs from EventSystem. Added Host and LocationID objects. Changed references from OriginatorID value to LocationID objects in CommandEvent and StatusEvent. Added references to Host and Location in eventsLib.i, and removed PipelineLogEvent.h Added unit tests for Host and LocationID
further notes about this change:

CommandEvent - Remove old references to OriginatorID.  These are replaced
with hostname and local id; these are used in conjunction with process id
to achieve the same purpose as OriginatorID, while fixing the issue of
the OriginatorID having a localID value that can potentially churn too
quickly.

The process ID in OriginatorID was originally 16-bits, but this broken on
Mac because it can have a larger than 16-bit pid.  This was temporarily
changed to 17-bits (to get a build running), and is now 32-bit with this
change.  Hostname was used to make it more user-friendly.  OriginatorID
could be addressed as one value before in an event selector; this can
still be achieved by using hostname, ipid, and localid together.  These
values can be retrieved together in a LocationID object.